implementing the interface segragation policy
