+++ app/src/main/java/com/example/clicker/di/HiltApplication.kt
@@ -9,6 +9,7 @@ import androidx.work.WorkerFactory
 import androidx.work.WorkerParameters
 import com.example.clicker.network.domain.TwitchRepo
 import com.example.clicker.data.workManager.OAuthTokeValidationWorker
+import com.example.clicker.network.domain.TwitchAuthentication
 import dagger.hilt.android.HiltAndroidApp
 import javax.inject.Inject
 
@@ -27,7 +28,7 @@ class HiltApplication:Application(), Configuration.Provider {
 }
 
 class CustomWorkerFactory @Inject constructor(
-    private val twitchRepoImpl: TwitchRepo
+    private val twitchRepoImpl: TwitchAuthentication
     ): WorkerFactory(){
     override fun createWorker(
         appContext: Context,
