+++ app/src/main/java/com/example/clicker/network/domain/TwitchAuthentication.kt
@@ -1,0 +1,11 @@
+package com.example.clicker.network.domain
+
+import com.example.clicker.network.models.ValidatedUser
+import com.example.clicker.util.Response
+import kotlinx.coroutines.flow.Flow
+
+interface TwitchAuthentication {
+    suspend fun validateToken(token:String): Flow<Response<ValidatedUser>>
+
+    fun logout(clientId:String,token:String):Flow<Response<String>>
+}
